{"schema":"decodex/commit/1","summary":"Split operator GitHub CLI authority status","authority":"XY-1115"}
